tests/requirements.txt
click==8.0.3
future==0.18.2
junitparser==2.2.0
PyYAML==6.0
termcolor==1.1.0
